refactor: moved functions to separate files
moved usePosts to hooks folder
