+++ packages/js/src/bulk-editor/helpers/index.js
@@ -1,0 +1,2 @@
+export { getSelectionView } from "./get-selection-view";
+export { getSmartSelectItems } from "./get-smart-select-items";
